% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{ad.nor}
\alias{ad.nor}
\title{Anderson-Darling Normality Test (ad.nor)}
\usage{
ad.nor(my.data)
}
\arguments{
\item{Numeric}{column position x in the dataframe}
}
\value{
P-value
}
\description{
Test for normality
}
\details{
Anderson-Darling Normality Test is used to determine whether a set of observations follows ‘Normal Distribution’.  The assumption of ‘Normality’ is widely used in Statistics in the areas of Inferential Statistics, Parametric Methods, Modeling (i.e.) from Simple Linear Regression to advanced Predictive Modeling Techniques.
}

