% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{kur}
\alias{kur}
\title{Kurtosis (kur)}
\usage{
kur(column.pos)
}
\arguments{
\item{Numeric}{(column position x in the dataframe)}
}
\value{
Gives the kurtosis value
}
\description{
Measure of Descriptive Statistics
}
\details{
Kurtosis addresses the characteristics of ‘Peakedness’ & ‘Taildness’.
}

