% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{ske}
\alias{ske}
\title{Skewness (ske)}
\usage{
ske(column.pos)
}
\arguments{
\item{Numeric}{column position x in the dataframe}
}
\value{
Gives the skewness value
}
\description{
Measure of Descriptive Statistics
}
\details{
Skewness addresses the characteristics of symmetry.
}

